pyproject.toml
[project]
name = "nhm-agent"
version = "1.0.0"
description = "Add your description here"
authors = [
    { name = "xuheng.zong", email = "79689669+nohairman@users.noreply.github.com" }
]
requires-python = ">=3.12,<3.15"
dependencies = [
    "dynaconf>=3.2.12",
    "fastapi>=0.127.0",
    "httpx>=0.28.1",
    "langchain>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "pydantic-settings>=2.12.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.4.11",
    "pre-commit>=4.5.1",
]

[tool.uv.sources]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true

[tool.black]
line_length = 120

[tool.isort]
profile = "black"
src_paths = ["src"]
py_version = 314
skip_gitignore = true
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
known_first_party = ["mercedes"]
multi_line_output = 3
line_length = 120
